refactor: changed Color to DeprecatedColor
