+++ components/ContactForm.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { DeprecatedColor } from '../Color.enum';
+import { DeprecatedColor } from '../DeprecatedColor.enum';
 import { Box, Text, Button } from 'rebass';
 import { Input, Textarea } from '@rebass/forms';
 import { themeDeprecated } from '../themeDeprecated';
